version: 1st ver of citybike data statistical analysis
Script prompt user to input how they want the data to be filtered and seen. Firstly, user chooses the city out of Chicago, New york and Washington. Secondly, user choose how they want the data to be filtered, by month, by day of the week, by both or by none at all. According to which we further ask for the month, the dow or both.